Added a separate file for cryptographic primitives
