+++ Client/Client.Core/Client.Core.Static/Client.cpp
@@ -5,8 +5,8 @@
 #include <Crypto.Static/Hashing.hpp>
 #include <Crypto.Static/ECDH.hpp>
 #include <Crypto.Static/AES_GCM.hpp>
-#include <Crypto.Static/RSAKeyGenerator.hpp>
 #include <Crypto.Static/RSA.hpp>
+#include <Crypto.Static/CryptoPrimitives.hpp>
 
 namespace Network
 {
@@ -212,8 +212,8 @@ void Client::userAuthorization(const std::string& login, const std::string& pass
 
     const std::string pwdHash       = Base::Hashing::SHA_256(password, login);
     const std::string verifyingHash = _connection->getConnectionVerifier()->calculateVerifyingHash(pwdHash, _connectionInfo);
-    const std::string encryptedVerifyingHash =
-        Base::Crypto::Asymmetric::RSA().encrypt(verifyingHash, Base::RSAKeyPair::getPublicKeyFromString(_connectionInfo._publicServerKey));
+    const std::string encryptedVerifyingHash = Base::Crypto::Asymmetric::RSA().encrypt(
+        verifyingHash, Base::Crypto::Primitives::RSAKeyPair::getPublicKeyFromString(_connectionInfo._publicServerKey));
 
     if (!encryptedVerifyingHash.empty())
     {
